backend/requirements.txt
pytapo
uvicorn
fastapi
requests
pymongo
python-dotenv
pydantic
opencv-python
aiofiles
pytest-asyncio
httpx
pytest
pytest-cov
pytest-mock
numpy
websockets